Implemented user management endpoints and authentication

+++ app/controllers/posts.server.controllers.js
@@ -1,5 +1,6 @@
 const posts = require("../models/posts.server.models")
 const Joi = require("joi")
+const users = require("../models/user.server.models");
 
 const add_post = (req, res) => {
     const schema = Joi.object({
@@ -9,7 +10,7 @@ const add_post = (req, res) => {
     const { error } = schema.validate(req.body);
     if(error) return res.sendStatus(400);
 
-    let post = Object.assign({}, req.body);
+    let post = Object.assign({}, req.body, {token:req.get("X-Authorization")});
 
     posts.addNewPost(post, (err, id) => {
         if (err) {
@@ -33,71 +34,93 @@ const get_post = (req, res) => {
 
 const update_post = (req, res) => {
     let post_id = parseInt(req.params.post_id);
+    let token = req.get("X-Authorization");
+
     posts.getSinglePost(post_id, (err, post) => {
         if(err === 404) return res.sendStatus(404);
         if(err) return res.sendStatus(500);
 
-        const schema = Joi.object({
-            "text": Joi.string().required()
-        });
-
-        const { error } = schema.validate(req.body);
-        if (error) return res.status(400).send(error.details[0].message);
-
-        if (post.text === req.body.text){
-            return res.sendStatus(200);
-        }
-
-        posts.updatePost(post_id, req.body.text, (err) => {
+        users.getIdFromToken(token, (err, id) => {
             if (err) return res.sendStatus(500);
-            return res.sendStatus(200);
-        })
 
+            if (id !== post.author.user_id) return res.sendStatus(403); // Check if the logged-in user is the author of the post
+
+            const schema = Joi.object({
+                "text": Joi.string().required()
+            });
+
+            const { error } = schema.validate(req.body);
+            if (error) return res.sendStatus(400);
+
+            if (post.text === req.body.text){
+                return res.sendStatus(200);
+            }
+
+            posts.updatePost(post_id, req.body.text, (err) => {
+                if (err) return res.sendStatus(500);
+                return res.sendStatus(200);
+            })
+        });
     });
 };
 
 const delete_post = (req, res) => {
     let post_id = parseInt(req.params.post_id);
-    posts.getSinglePost(post_id, (err) => {
+    let token = req.get("X-Authorization");
+
+    posts.getSinglePost(post_id, (err, post) => {
         if(err === 404) return res.sendStatus(404);
         if(err) return res.sendStatus(500);
 
-        // check auth
-
-        posts.deletePost(post_id, (err) => {
+        users.getIdFromToken(token, (err, id) => {
             if (err) return res.sendStatus(500);
-            return res.sendStatus(200);
-        });
 
+            if (id !== post.author.user_id) return res.sendStatus(403); // Check if the logged-in user is the author of the post
+
+            posts.deletePost(post_id, (err) => {
+                if (err) return res.sendStatus(500);
+                return res.sendStatus(200);
+            });
+        });
     });
 };
 
 const add_like = (req, res) => {
     let post_id = parseInt(req.params.post_id);
+    let token = req.get("X-Authorization");
 
     posts.getSinglePost(post_id, (err) => {
         if(err === 404) return res.sendStatus(404);
         if(err) return res.sendStatus(500);
 
-        posts.addLike(post_id, (err) => {
-            if(err === 403) return res.sendStatus(403);
+        users.getIdFromToken(token, (err, id) => {
             if (err) return res.sendStatus(500);
-            return res.sendStatus(200);
+
+            posts.addLike(post_id, id, (err) => {
+                if(err === 403) return res.sendStatus(403);
+                if (err) return res.sendStatus(500);
+                return res.sendStatus(200);
+            });
         });
     });
 };
 
 const remove_like = (req, res) => {
     let post_id = parseInt(req.params.post_id);
+    let token = req.get("X-Authorization");
 
     posts.getSinglePost(post_id, (err) => {
         if(err === 404) return res.sendStatus(404);
         if(err) return res.sendStatus(500);
 
-        posts.removeLike(post_id, (err) => {
-            if(err === 403) return res.sendStatus(403);
+        users.getIdFromToken(token, (err, id) => {
             if (err) return res.sendStatus(500);
-            return res.sendStatus(200);
+
+            posts.removeLike(post_id, id, (err) => {
+                if(err === 403) return res.sendStatus(403);
+                if (err) return res.sendStatus(500);
+                return res.sendStatus(200);
+            });
         });
     });
